fix develop (その他のコンポーネントも分離)
